Исправлены ошибки разграничения уровней абстракции.
